% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inertiafunctions.R
\name{calc_inertia_conesolid}
\alias{calc_inertia_conesolid}
\title{Moment of inertia tensor of a solid circular cone pyramid}
\usage{
calc_inertia_conesolid(r, h, m)
}
\arguments{
\item{r}{radius of the cone base (m)}

\item{h}{height of the cone (m)}

\item{m}{mass of the cone (kg)}
}
\value{
a 3x3 matrix representing the moment of inertia tensor of a solid
circular cone about
its center of gravity with z oriented through it's major axis.
}
\description{
All outputs are based on an origin at the centered point on the base
Reference: https://apps.dtic.mil/sti/pdfs/AD0274936.pdf
}
\section{CAUTION}{
 Origin of the output tensor is NOT at the center of gravity but
at the center of the base.
}

\author{
Christina Harvey
}
