% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mathematicalfunctions.R
\name{calc_inertia_cylsolid}
\alias{calc_inertia_cylsolid}
\title{Moment of inertia tensor of a solid cylinder}
\usage{
calc_inertia_cylsolid(r, h, m)
}
\arguments{
\item{r}{radius of the cylinder}

\item{h}{height of the cylinder}

\item{m}{mass of the cylinder}
}
\value{
Function returns the moment of inertia tensor of a solid cylinder about
its center of gravity with z oriented through it's major axis
}
\description{
Moment of inertia tensor of a solid cylinder
}
\examples{

}
\author{
Christina Harvey
}
