% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inertiafunctions.R
\name{calc_inertia_ellcone}
\alias{calc_inertia_ellcone}
\title{Moment of inertia tensor of a solid elliptical cone - end of purple notebook
derivation verified in green}
\usage{
calc_inertia_ellcone(A, B, l, m)
}
\arguments{
\item{A}{half height of the wide base (m)}

\item{B}{half width of the wide base (m)}

\item{l}{length of the cone (m)}

\item{m}{mass of the cone (kg)}
}
\value{
a 3x3 matrix representing the moment of inertia tensor of a solid
elliptical cone about the center of the wider base with z oriented
towards the end.
}
\description{
Moment of inertia tensor of a solid elliptical cone - end of purple notebook
derivation verified in green
}
\section{CAUTION}{
 Origin of the output tensor is NOT at the center of gravity
but at the center of the base.
}

\author{
Christina Harvey
}
