% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inertiafunctions.R
\name{calc_inertia_ellcyl}
\alias{calc_inertia_ellcyl}
\title{Moment of inertia tensor of a solid elliptical cylinder
Reference: https://apps.dtic.mil/sti/pdfs/AD0274936.pdf}
\usage{
calc_inertia_ellcyl(a, b, l, m)
}
\arguments{
\item{a}{half height of the base - oriented along the x axis in torso FOR
(z axis in full bird FOR) (m)}

\item{b}{half width of the base - oriented along the y axis in torso FOR
(y axis in full bird FOR)  (m)}

\item{l}{length of the cylinder - oriented along the z axis in torso FOR
(x axis in full bird FOR)  (m)}

\item{m}{mass of the cylinder (kg)}
}
\value{
a 3x3 matrix representing the moment of inertia tensor of a solid
elliptical cylinder about it's center of gravity
}
\description{
Moment of inertia tensor of a solid elliptical cylinder
Reference: https://apps.dtic.mil/sti/pdfs/AD0274936.pdf
}
\author{
Christina Harvey
}
