% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inertiafunctions.R
\name{calc_inertia_pyrasolid}
\alias{calc_inertia_pyrasolid}
\title{Moment of inertia tensor of a solid square pyramid
Reference: https://apps.dtic.mil/sti/pdfs/AD0274936.pdf
All outputs are based on an origin at the centered point on the base}
\usage{
calc_inertia_pyrasolid(w, h, m)
}
\arguments{
\item{w}{entire width of one side of the pyramid base (m)}

\item{h}{entire height of the pyramid (m)}

\item{m}{mass of the pyramid (kg)}
}
\value{
a 3x3 matrix representing the moment of inertia tensor of a solid
square pyramid about its center of gravity with z oriented through it's
major axis.
}
\description{
Moment of inertia tensor of a solid square pyramid
Reference: https://apps.dtic.mil/sti/pdfs/AD0274936.pdf
All outputs are based on an origin at the centered point on the base
}
\section{CAUTION}{
 Origin is NOT at the center of gravity but at the center of the base.
}

\author{
Christina Harvey
}
