% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{clean_pts}
\alias{clean_pts}
\title{The identified peripheral and joint 3D positions.}
\format{
A matrix with 10 rows and 3 columns. Columns represent x, y, and z coordinate respectively:
\describe{
\item{pt1x, pt1y, pt1z}{Point on the shoulder joint (m).}
\item{pt2x, pt1y, pt2z}{Point on the elbow joint (m).}
\item{pt3x, pt3y, pt3z}{Point on the wrist joint (m).}
\item{pt4x, pt4y, pt4z}{Point on the end of carpometacarpus (m).}
\item{pt6x, pt6y, pt6z}{Point on the leading edge of the wing in front of the
wrist joint (m).}
\item{pt8x, pt8y, pt8z}{Point on tip of most distal primary (m).}
\item{pt9x, pt9y, pt9z}{Point on the tip of the last primary to model as if
it is on the end of the carpometacarpus (m).}
\item{pt10x, pt10y, pt10z}{Point on tip of last primary to model as if
it was distributed along the carpometacarpus (m).}
\item{pt11x, pt11y, pt11z}{Point on tip of most proximal feather (m).}
\item{pt12x, pt12y, pt12z}{Point on exterior shoulder position
(wing root leading edge) (m).}
}
}
\usage{
clean_pts
}
\description{
A dataset containing the identified peripheral and joint 3D positions for a
pigeon BirdID = 20_0300, TestID = 20_03004, FrameID = 317
}
\keyword{datasets}
