% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{dat_bone_curr}
\alias{dat_bone_curr}
\title{Wing bone specific morphology dataset}
\format{
A dataframe with 6 rows and 5 variables:
\describe{
\item{bone}{Bone ID code. Must include:
"Humerus","Ulna","Radius","Carpometacarpus","Ulnare" and "Radiale".}
\item{bone_mass}{Mass of bone in the same row as the appropriate
bone ID code (kg)}
\item{bone_len}{Length of bone in the same row as the appropriate
bone ID code (m)}
\item{bone_out_rad}{Outer radius of bone in the same row as the appropriate
bone ID code (m)}
\item{bone_in_rad}{Inner radius of bone in the same row as the appropriate
bone ID code (m)}
}
}
\usage{
dat_bone_curr
}
\description{
A dataset containing the wing bone specific data for a
pigeon BirdID = 20_0300, TestID = 20_03004, FrameID = 317
}
\keyword{datasets}
