% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{dat_feat_curr}
\alias{dat_feat_curr}
\title{Flight feather specific morphology dataset}
\format{
A dataframe with 20 rows and 15 variables.
Only 8 variables required for proper functioning:
\describe{
\item{feather}{Feather ID code. Must be in standard format i.e.
1st primary is "P1", third secondary is "S3", etc.
Alula feathers should be grouped and named "alula".}
\item{m_f}{Mass of feather in the same row as the
appropriate feather ID code (kg)}
\item{l_cal}{Length of calamus in the same row as the
appropriate feather ID code (m)}
\item{l_vane}{Length of rachis/vane in the same row as the
appropriate feather ID code (m)}
\item{w_cal}{Width (diameter) of calamus in the same row as the
appropriate feather ID code (m)}
\item{w_vp}{Width of proximal vane (average value) in the same row as the
appropriate feather ID code (m)}
\item{w_vd}{Width of distal vane (average value)  in the same row as the
appropriate feather ID code (m)}
\item{vane_angle}{Interior angle between the rachis and calamus  (degrees)}
}
}
\usage{
dat_feat_curr
}
\description{
A dataset containing the wing bone specific data for a
pigeon BirdID = 20_0300, TestID = 20_03004, FrameID = 317
}
\keyword{datasets}
