% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{dat_id_curr}
\alias{dat_id_curr}
\title{Identification variables for current configuration}
\format{
A dataframe with 1 row and 4 required variables:
\describe{
\item{species}{Species ID code as a string}
\item{BirdID}{Bird ID code as a string}
\item{TestID}{Test ID code as a string}
\item{frameID}{Video frame ID code as a string}
}
}
\usage{
dat_id_curr
}
\description{
A dataset containing the identification data for a
pigeon BirdID = 20_0300, TestID = 20_03004, FrameID = 317
}
\keyword{datasets}
