% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{dat_mat}
\alias{dat_mat}
\title{Material properties.}
\format{
A dataframe with 20 rows and 15 variables.
Only 8 variables required for proper functioning:
\describe{
\item{material}{Material information. Must include the following:
"Bone","Skin","Muscle","Cortex", "Medullary"}
\item{density}{Density of each material (kg/m^3)}
}
}
\usage{
dat_mat
}
\description{
A dataset containing the material properties for a
pigeon BirdID = 20_0300, TestID = 20_03004, FrameID = 317.
}
\keyword{datasets}
