% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mathematicalfunctions.R
\name{dirac_delta}
\alias{dirac_delta}
\title{Dirac delta function}
\usage{
dirac_delta(i, j)
}
\arguments{
\item{i}{an index (usually row of a matrix)}

\item{j}{an index (usually column of a matrix)}
}
\value{
If i and j are equal dirac returns 1 otherwise returns 0
}
\description{
Dirac delta function
}
\examples{

 # should return 1
 dirac_delta(7,7)

 # should return 0
 dirac_delta(5,4)

}
\author{
Christina Harvey
}
