% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mathematicalfunctions.R
\name{kronecker_delta}
\alias{kronecker_delta}
\title{Kroneckerdelta function}
\usage{
kronecker_delta(i, j)
}
\arguments{
\item{i}{a scalar index (usually row of a matrix)}

\item{j}{a scalar index (usually column of a matrix)}
}
\value{
a scalar value. Returns 1 if i and j are equal otherwise returns 0
}
\description{
Kroneckerdelta function
}
\examples{

library(AvInertia)

# should return 1
kronecker_delta(1,1)

# should return 0
kronecker_delta(0,1)

}
\author{
Christina Harvey
}
