% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/anatomicalmassprop.R
\name{massprop_muscles}
\alias{massprop_muscles}
\title{Muscle mass properties}
\usage{
massprop_muscles(m, rho, l, start, end)
}
\arguments{
\item{m}{Mass of muscle (kg)}

\item{rho}{Density of muscle (kg/m^3)}

\item{start}{a 1x3 vector (x,y,z) representing the 3D point where bone starts.
Frame of reference: VRP | Origin: VRP}

\item{end}{a 1x3 vector (x,y,z) representing the 3D point where bone ends.
Frame of reference: VRP | Origin: VRP}
}
\value{
This function returns a list that includes:
\itemize{
\item{I}{a 3x3 matrix representing the moment of inertia tensor of a muscle
modeled as a solid cylinder distributed along the bone length}
\item{CG}{a 1x3 vector representing the center of gravity position of a muscle
modeled as a solid cylinder distributed along the bone length}
\item{m}{a double that returns the input muscle mass}
}
}
\description{
Calculate the moment of inertia of a muscle modeled as a solid cylinder
distributed along the bone length
}
\section{Warning}{

Parallel axis theorem does not apply between two arbitrary points. One point
must be the object's center of gravity.
}

\author{
Christina Harvey
}
