% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/anatomicalmassprop.R
\name{massprop_pm}
\alias{massprop_pm}
\title{Calculate the moment of inertia of of any point mass}
\usage{
massprop_pm(m, pt)
}
\arguments{
\item{m}{mass of skin (kg)}

\item{pt}{three points defining the location of the point mass. Frame of reference: VRP | Origin: VRP
each point should be a different row as follows:
pt1x, pt1y, pt1z}
}
\value{

}
\description{
Calculate the moment of inertia of of any point mass
}
\author{
Christina Harvey
}
