% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mainrunfunctions.R
\name{massprop_restbody}
\alias{massprop_restbody}
\title{Calculate the center of gravity and moment of inertia for the head, neck,
torso and tail}
\usage{
massprop_restbody(dat_wingID_curr, dat_bird_curr)
}
\arguments{
\item{dat_wingID_curr}{Dataframe related to the current bird wing ID
info that must include the following columns:
\itemize{
\item{species}{Species ID code as a string}
\item{BirdID}{Bird ID code as a string}
\item{TestID}{Test ID code as a string}
\item{frameID}{Video frame ID code as a string}
}}

\item{dat_bird_curr}{Dataframe related to the current bird wing that must
include the following columns:
\itemize{
\item{extend_neck}{Logical input defining whether the neck should be
modelled as extended or not}
\item{head_length}{Mass of full bird for the current wing (kg)}
\item{head_mass}{Mass of one wing, should be the current wing (kg)}
\item{head_height}{Radius of feather barb  for current species (m)}
\item{neck_mass}{Distance between feather barbs for current species (m)}
\item{neck_width}{Mass of all muscles in the brachial region
of the wing (kg)}
\item{neck_length}{Mass of all muscles in the antebrachial region
of the wing (kg)}
\item{torsotail_length}{Mass of all muscles in the manus region
of the wing (kg)}
\item{torsotail_mass}
\item{tail_length}
\item{tail_mass}
\item{tail_width}
\item{right_leg_mass}
\item{left_leg_mass}
\item{body_width_max}
\item{body_width_at_leg_insert}
\item{x_loc_of_body_max}
\item{x_loc_leg_insertion}
\item{x_loc_TorsotailCoG}
\item{z_loc_TorsotailCoG}
}}
}
\value{
Function returns a dataframe that includes the moment of inertia and
center of gravity of head, neck, torso and tail
}
\description{
Function that reads in anatomical data and returns the moment of inertia
tensor and center of gravity for the head, neck, tail and torso
}
