% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mathematicalfunctions.R
\name{rotx}
\alias{rotx}
\title{x-axis rotation matrix}
\usage{
rotx(angle)
}
\arguments{
\item{angle}{angle to rotate}
}
\value{

}
\description{
x-axis rotation matrix
}
\examples{

}
\author{
Christina Harvey
}
