% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inertiafunctions.R
\name{calc_inertia_cylsolid}
\alias{calc_inertia_cylsolid}
\title{Moment of inertia tensor of a solid cylinder}
\usage{
calc_inertia_cylsolid(r, h, m)
}
\arguments{
\item{r}{radius of the cylinder (m)}

\item{h}{height of the cylinder (m)}

\item{m}{mass of the cylinder (kg)}
}
\value{
a 3x3 matrix representing the moment of inertia tensor of
a solid cylinder about its center of gravity with z oriented
through it's major axis
}
\description{
Reference: https://apps.dtic.mil/sti/pdfs/AD0274936.pdf
}
\author{
Christina Harvey
}
