% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inertiafunctions.R
\name{calc_inertia_ellipse}
\alias{calc_inertia_ellipse}
\title{Moment of inertia tensor of solid ellipse CG or a half ellipse centered on
the base}
\usage{
calc_inertia_ellipse(a, b, c, m)
}
\arguments{
\item{a}{half the height along the x direction (m)}

\item{b}{half the width along the y direction  (m)}

\item{c}{half the length along the z direction (m)}

\item{m}{mass of the ellipse (kg)}
}
\value{
a 3x3 matrix representing the moment of inertia tensor of a solid
ellipse about its center of gravity with the major axes aligned.
}
\description{
Reference:https://apps.dtic.mil/sti/pdfs/AD0274936.pdf
}
\section{CAUTION}{
 Origin is at the center of gravity for a full ellipse or at the
center of the base if modelling a half ellipse.
}

\author{
Christina Harvey
}
