% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inertiafunctions.R
\name{calc_inertia_platerect}
\alias{calc_inertia_platerect}
\title{Moment of inertia tensor of a flat rectangular plate - assumes thickness is
approximately zero
Reference: https://apps.dtic.mil/sti/pdfs/AD0274936.pdf}
\usage{
calc_inertia_platerect(w, h, m)
}
\arguments{
\item{w}{full width of one side of the plate - short edge (m)}

\item{h}{height of the plate - long edge (m)}

\item{m}{mass of the plate (kg)}
}
\value{
a 3x3 matrix representing the moment of inertia tensor of a flat
plate about its center of gravity with z oriented parallel with it's long
edge (h) and y along its short edge (w)
}
\description{
Moment of inertia tensor of a flat rectangular plate - assumes thickness is
approximately zero
Reference: https://apps.dtic.mil/sti/pdfs/AD0274936.pdf
}
\author{
Christina Harvey
}
