% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mathematicalfunctions.R
\name{calc_inertia_platerect}
\alias{calc_inertia_platerect}
\title{Moment of inertia tensor of a solid square pyramid}
\usage{
calc_inertia_platerect(w, h, m)
}
\arguments{
\item{w}{full width of one side of the plate}

\item{h}{height of the plate}

\item{m}{mass of the plate}
}
\value{
Function returns the moment of inertia tensor of a flat plate about
its center of gravity with z oriented parallel with it's long edge
}
\description{
Moment of inertia tensor of a solid square pyramid
}
\examples{

}
\author{
Christina Harvey
}
