% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inertiafunctions.R
\name{calc_inertia_platetri}
\alias{calc_inertia_platetri}
\title{Moment of inertia tensor of a flat triangular plate}
\usage{
calc_inertia_platetri(pts, A, rho, t, desired_prop)
}
\arguments{
\item{pts}{a matrix of the three 3D points that define a point on
the triangular plate.
Frame of reference: Muscle | Origin: VRP
each point should be a different row as follows:
pt1x, pt1y, pt1z
pt2x, pt1y, pt2z
pt3x, pt3y, pt3z}

\item{A}{area of the triangular plate (m)}

\item{rho}{density of the material (kg/m^3)}

\item{t}{thickness of the plate (m)}

\item{desired_prop}{a string containing either "I" or "CG" depending on the
desired output}
}
\value{
a 3x3 matrix representing the moment of inertia tensor of a flat
triangular plate about
its center of gravity. Z axis defined as the normal to the input points.
}
\description{
Reference: https://apps.dtic.mil/dtic/tr/fulltext/u2/a183444.pdf
page 4 equations 2.16-2.20
}
\section{Warning}{

The input points should be defined in a counterclockwise direction around the
plate in the triangular plate frame of reference
i.e. all z components should be equal
}

\author{
Christina Harvey
}
