% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mathematicalfunctions.R
\name{calc_inertia_pyrasolid}
\alias{calc_inertia_pyrasolid}
\title{Moment of inertia tensor of a solid square pyramid about the centered point on the base}
\usage{
calc_inertia_pyrasolid(w_2, h, m)
}
\arguments{
\item{w_2}{half width of one side of the pyramid base}

\item{h}{height of the pyramid}

\item{m}{mass of the pyramid}
}
\value{
Function returns the moment of inertia tensor of a solid square pyramid cylinder about
its center of gravity with z oriented through it's major axis
}
\description{
Moment of inertia tensor of a solid square pyramid about the centered point on the base
}
\examples{

}
\author{
Christina Harvey
}
