% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mathematicalfunctions.R
\name{calc_rot}
\alias{calc_rot}
\title{Title}
\usage{
calc_rot(z_vector, x_vector)
}
\arguments{
\item{x_vector}{}
}
\value{
Function returns the cosine rotation matrices to transform between VRP frame and object frame
}
\description{
Title
}
\examples{

}
\author{
Christina Harvey
}
