% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mathematicalfunctions.R
\name{calc_univec}
\alias{calc_univec}
\title{Determine the unit vector of any input vector}
\usage{
calc_univec(vector)
}
\arguments{
\item{vector}{any vector or array with only one dimension}
}
\value{
the unit vector in the size of the input vector
}
\description{
Determine the unit vector of any input vector
}
\author{
Christina Harvey
}
