% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mathematicalfunctions.R
\name{calc_univec}
\alias{calc_univec}
\title{Determine the unit vector of any input vector}
\usage{
calc_univec(vector)
}
\arguments{
\item{vector}{}
}
\value{
This function returns the unit vector of any input vector
}
\description{
Determine the unit vector of any input vector
}
\examples{

output = unit_vector(c(2,6,5))
#should equal one if a unit vector
length_output = pracma::Norm(output)

}
\author{
Christina Harvey
}
