% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mainrunfunctions.R
\name{combine_inertialprop}
\alias{combine_inertialprop}
\title{Combine body and wing inertial components.}
\usage{
combine_inertialprop(
  curr_torsotail_data,
  left_wing_data,
  right_wing_data,
  symmetric
)
}
\arguments{
\item{curr_torsotail_data}{{Output dataframe from massprop_restbody}}

\item{left_wing_data}{{Output dataframe from massprop_birdwing}}

\item{right_wing_data}{{Output dataframe from massprop_birdwing}}

\item{symmetric}{{Logical indicating if the input wings are symmetric or not.
If True than left_wing_data = right_wing_data}}
}
\description{
Combines data exported from massprop_restbody
and massprop_birdwing.
}
