% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mainrunfunctions.R
\name{massprop_birdwing}
\alias{massprop_birdwing}
\title{Function that reads in anatomical data and returns the moment of inertia tensor and center
of gravity of a wing one side of the bird}
\usage{
massprop_birdwing(
  dat_bird_curr,
  dat_bone_curr,
  dat_feat_curr,
  dat_mat_curr,
  clean_pts
)
}
\arguments{
\item{clean_pts}{data frame of the key positions of the bird as follows:
pt1x, pt1y, pt1z - Point that defines the shoulder joint
pt2x, pt1y, pt2z - Point that defines the elbow joint
pt3x, pt3y, pt3z - Point that defines the wrist joint
pt4x, pt4y, pt4z - Point that defines the end of carpometacarpus}
}
\value{

}
\description{
Function that reads in anatomical data and returns the moment of inertia tensor and center
of gravity of a wing one side of the bird
}
