% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/anatomicalmassprop.R
\name{massprop_bones}
\alias{massprop_bones}
\title{Calculate the moment of inertia of a bone modelled as a hollow cylinder with two solid end caps}
\usage{
massprop_bones(m, l, r_out, r_in, rho, start, end)
}
\arguments{
\item{m}{full mass of the bone}

\item{l}{full length of the bone}

\item{r_out}{outer radius of bone}

\item{r_in}{inner radius of bone}

\item{rho}{density of the bone (kg/m^3)}

\item{start}{3D point where bone starts. Frame of reference: VRP | Origin: VRP}

\item{end}{3D point where bone ends. Frame of reference: VRP | Origin: VRP}
}
\value{

}
\description{
Calculate the moment of inertia of a bone modelled as a hollow cylinder with two solid end caps
}
\section{Warning}{

the parallel axis theorem only works if the moment of inertia is calculated about components centroidal axis
}

