% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/anatomicalmassprop.R
\name{massprop_feathers}
\alias{massprop_feathers}
\title{Calculate the moment of inertia of skin modelled as a flat triangular plate}
\usage{
massprop_feathers(
  m_f,
  l_c,
  l_r_cor,
  w_r,
  r_b,
  d_b,
  rho_cor,
  rho_med,
  w_vp,
  w_vd,
  angle,
  normal,
  start,
  end
)
}
\arguments{
\item{m_f}{mass of the entire feather}

\item{l_c}{length of the calamus - start of vane to end of calamus(m)}

\item{l_r_cor}{length of rachis/vane - tip to start of vane (m)}

\item{w_r}{width (diameter) of the cortex part of the calamus (m)}

\item{r_b}{barb radius (m)}

\item{d_b}{barb distance (m)}

\item{rho_cor}{density of the cortex (kg/m^3)}

\item{rho_med}{density of the medullary (kg/m^3)}

\item{w_vp}{width of proximal (closest to body) vane (m)}

\item{w_vd}{width of distal (closest to wing tip) vane (m)}

\item{angle}{angle between l_c and l_r}

\item{start}{3D point where feather starts. Frame of reference: VRP | Origin: VRP}

\item{end}{3D point where feather ends Frame of reference: VRP | Origin: VRP}

\item{n_pts}{the pts that define the plane that the feather lies flat on.
Order counterclockwise so that the normal will pointing downwards from the ventral surface of the feather.
Frame of reference: VRP | Origin: VRP}
}
\value{

}
\description{
Calculate the moment of inertia of skin modelled as a flat triangular plate
}
\section{Warning}{

Parallel axis theorem is only valid if relocating an object between it's CG and an arbitrary pt. I_arbitrary = I_CG + md^2
Cannot be used between two arbitrary points.
}

\author{
Christina Harvey
}
