% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/anatomicalmassprop.R
\name{massprop_feathers}
\alias{massprop_feathers}
\title{Feather mass properties}
\usage{
massprop_feathers(
  m_f,
  l_c,
  l_r_cor,
  w_cal,
  r_b,
  d_b,
  rho_cor,
  rho_med,
  w_vp,
  w_vd,
  angle
)
}
\arguments{
\item{m_f}{Mass of the entire feather (kg)}

\item{l_c}{Length of the calamus; start of vane to end of calamus(m)}

\item{l_r_cor}{Length of rachis; tip to start of vane (m)}

\item{w_cal}{Width (diameter) of the cortex part of the calamus (m)}

\item{r_b}{Radius of feather barbs (m)}

\item{d_b}{Distance between barbs (m)}

\item{rho_cor}{Density of the cortex (kg/m^3)}

\item{rho_med}{Density of the medullary (kg/m^3)}

\item{w_vp}{Width of proximal (closest to body) vane (m)}

\item{w_vd}{Width of distal (closest to wing tip) vane (m)}

\item{angle}{Angle between calamus and the vane taken the supplement angle
to the interior angle. Negative indicates the feather tip is rotated
proximally relative to the start of the feather vane.}
}
\value{
This function returns a list that includes:
\itemize{
\item{I}{a 3x3 matrix representing the moment of inertia tensor of a
simplified feather}
\item{CG}{a 1x3 vector representing the center of gravity position of a
simplified feather}
\item{m}{a double that returns the feather mass}
}
}
\description{
Calculate the moment of inertia of skin modeled as a flat triangular plate
}
\section{Warning}{

Parallel axis theorem does not apply between two arbitrary points.
One point must be the object's center of gravity.

CAUTION: While computing the variable components of the feather the x axis
is the normal of the feather.
}

\author{
Christina Harvey
}
