% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/anatomicalmassprop.R
\name{massprop_muscles}
\alias{massprop_muscles}
\title{Calculate the moment of inertia of a muscle modelled as a solid cylinder distributed along the bone length}
\usage{
massprop_muscles(m, rho, start, end)
}
\arguments{
\item{m}{mass of muscle (kg)}

\item{rho}{density of muscle (kg/m^3)}

\item{start}{3D point where bone starts. Frame of reference: VRP | Origin: VRP}

\item{end}{3D point where bone ends. Frame of reference: VRP | Origin: VRP}
}
\value{

}
\description{
Calculate the moment of inertia of a muscle modelled as a solid cylinder distributed along the bone length
}
