% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/anatomicalmassprop.R
\name{massprop_pm}
\alias{massprop_pm}
\title{Point-mass mass properties}
\usage{
massprop_pm(m, pt)
}
\arguments{
\item{m}{Mass of skin (kg)}

\item{pt}{a 1x3 vector (x,y,z) representing the  location of the point mass. Frame of reference: VRP | Origin: VRP}
}
\value{
This function returns a list that includes:
\itemize{
\item{I}{a 3x3 matrix representing the moment of inertia tensor of a point mass}
\item{CG}{a 1x3 vector representing the center of gravity position of a point mass}
}
}
\description{
Calculate the moment of inertia of any point mass
}
\section{Warning}{

Parallel axis theorem does not apply between two arbitrary points. One point must be the object's center of gravity.
}

\author{
Christina Harvey
}
