% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/anatomicalmassprop.R
\name{massprop_skin}
\alias{massprop_skin}
\title{Calculate the moment of inertia of skin modelled as a flat triangular plate}
\usage{
massprop_skin(m, rho, pts)
}
\arguments{
\item{m}{mass of skin (kg)}

\item{rho}{density of skin (kg/m^3)}

\item{pts}{three points defining the vertices of the triangle. Frame of reference: VRP | Origin: VRP
Must be numbered in a counterclockwise direction for positive area, otherwise signs will be reversed.
each point should be a different row as follows:
pt1x, pt1y, pt1z
pt2x, pt1y, pt2z
pt3x, pt3y, pt3z
For inner skin: row1 - Shoulder, row2 - Wrist, row3 - Elbow * ensures CCW ordering
For outer skin: row1 - Wrist, row2 - Finger, row3 - Elbow * ensures CCW ordering}
}
\value{

}
\description{
Calculate the moment of inertia of skin modelled as a flat triangular plate
}
\author{
Christina Harvey
}
