% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mathematicalfunctions.R
\name{parallelaxis}
\alias{parallelaxis}
\title{Reads in an initial tensor and an offset to compute the transformed tensor.
Will be in the same frame of reference as the input tensor.}
\usage{
parallelaxis(I_CG, offset_vec, m)
}
\arguments{
\item{m}{}
}
\value{
Function returns the transformed tensor after a solid body translation
}
\description{
Reads in an initial tensor and an offset to compute the transformed tensor.
Will be in the same frame of reference as the input tensor.
}
\examples{

}
\author{
Christina Harvey
}
