% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mathematicalfunctions.R
\name{rotx}
\alias{rotx}
\title{A 3x3 rotation matrix allowing rotation about the x-axis.
Constructed using a cosine rotation matrix where the rotation angle in
degrees is measured counterclockwise allowing positive rotation under
the right hand rule.}
\usage{
rotx(angle)
}
\arguments{
\item{angle}{a scalar representing the angle to rotate (degrees)}
}
\value{
a 3x3 matrix representing the rotation about the x-axis by the
given angle
}
\description{
A 3x3 rotation matrix allowing rotation about the x-axis.
Constructed using a cosine rotation matrix where the rotation angle in
degrees is measured counterclockwise allowing positive rotation under
the right hand rule.
}
\examples{

library(AvInertia)
angle = 90
# should return matrix [[1,0,0];[0,0,1];[0,1,0]]
rotx(angle)

}
\author{
Christina Harvey
}
