% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mainrunfunctions.R
\name{store_data}
\alias{store_data}
\title{Function to store moment of inertia tensor and center of gravity vector components in long format}
\usage{
store_data(species_curr, alldat_row, dat_mass, mass_properties, name)
}
\arguments{
\item{species_curr}{current species code}

\item{alldat_row}{current dataframe containing the necessary identifying information for the wing shape}

\item{dat_mass}{new MOI and CG data to add to mass_properties as new rows}

\item{mass_properties}{current master data fram that saves the data}

\item{name}{name of the component}
}
\value{

}
\description{
Function to store moment of inertia tensor and center of gravity vector components in long format
}
