% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mainrunfunctions.R
\name{store_data}
\alias{store_data}
\title{Store data from the inertia calculations in long format}
\usage{
store_data(dat_wingID_curr, dat_mass, mass_properties, name)
}
\arguments{
\item{dat_wingID_curr}{Dataframe related to the current bird wing ID
info that must include the following columns:
\itemize{
\item{species}{Species ID code as a string}
\item{BirdID}{Bird ID code as a string}
\item{TestID}{Test ID code as a string}
\item{frameID}{Video frame ID code as a string}
}}

\item{dat_mass}{Dataframe containing the new MOI and CG data to add
to mass_properties as new rows. Must include:
\itemize{
\item{I}{Moment of inertia tensor (kg-m^2)}
\item{CG}{Center of gravity with three location components (m)}
}}

\item{mass_properties}{Dataframe containing any previously saved data.
Must have the following columns: "species","BirdID","TestID","FrameID",
"prop_type","component","value".}

\item{name}{Name of the component for which the moment of inertia and
center of gravity were computed.}
}
\value{
This function returns mass_properties as an updated dataframe
with a new row corresponding to the dat_mass information
}
\description{
Function to store moment of inertia tensor and center of gravity vector
components in long format
}
